+++ corpus/invalid_symbols/foo.clj
@@ -1,0 +1,2 @@
+'foo/bar/baz
+#_:clj-kondo/ignore 'foo/bar/baz